[SD] Complete adapters/ work and build
